Refactor shell service tests
