+++ misc/ls-extensions/modules/bal-shell-service/src/test/resources/testcases/getResult/basic_variables.json
