+++ misc/ls-extensions/modules/bal-shell-service/src/test/resources/testcases/getResult/mime_types.json
