+++ misc/ls-extensions/modules/bal-shell-service/src/test/resources/testcases/getResult/simple_arithmetic.json
